Make sure to never delete the default stream in StreamFacade#delete()
The default stream might be included in a content pack but it should
never be deleted when a content pack gets uninstalled.

+++ graylog2-server/src/main/java/org/graylog2/contentpacks/facades/StreamFacade.java
@@ -225,6 +225,10 @@ public class StreamFacade implements EntityFacade<Stream> {
 
     @Override
     public void delete(Stream nativeEntity) {
+        if (nativeEntity.isDefaultStream()) {
+            LOG.debug("The default stream should not be deleted");
+            return;
+        }
         try {
             streamService.destroy(nativeEntity);
         } catch (NotFoundException ignore) {
